requirements.txt
numpy==1.21.6
pandas==1.1.5
python-dotenv==0.21.1
scikit-learn==1.0.2
spotipy==2.23.0
streamlit==1.23.1
urllib3==1.26.6
